Cleanup/Refactor/CodeChange (Broken maybe)
Cleaned up a lot of my code, refactored a lot of my code and cleaned up
most things. Changed up everything to handle the new formatting, removed
some sample code and cleaned project files up. There is a lot more that
needs to be done this is just a major refactor push